Refactor quote extraction service boundary
