+++ backend/app/features/quotes/service.py
@@ -3,8 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-from collections.abc import Sequence
-from dataclasses import dataclass
 from datetime import UTC, datetime
 from typing import Protocol, cast
 from uuid import UUID, uuid4
@@ -20,15 +18,11 @@ from app.features.quotes.repository import (
     QuoteRenderContext,
 )
 from app.features.quotes.schemas import (
-    ExtractionResult,
     LineItemDraft,
     QuoteCreateRequest,
     QuoteUpdateRequest,
 )
-from app.integrations.audio import AudioClip, AudioError
-from app.integrations.extraction import ExtractionError
 from app.integrations.pdf import PdfRenderError
-from app.integrations.transcription import TranscriptionError
 
 
 class QuoteServiceError(Exception):
@@ -92,138 +86,23 @@ class QuoteRepositoryProtocol(Protocol):
     async def rollback(self) -> None: ...
 
 
-class ExtractionIntegrationProtocol(Protocol):
-    """Structural protocol for extraction integration dependency."""
-
-    async def extract(self, notes: str) -> ExtractionResult: ...
-
-
 class PdfIntegrationProtocol(Protocol):
     """Structural protocol for PDF rendering integration dependency."""
 
     def render(self, context: QuoteRenderContext) -> bytes: ...
 
 
-class AudioIntegrationProtocol(Protocol):
-    """Structural protocol for audio normalization integration dependency."""
-
-    def normalize_and_stitch(self, clips: Sequence[AudioClip]) -> bytes: ...
-
-
-class TranscriptionIntegrationProtocol(Protocol):
-    """Structural protocol for speech-to-text integration dependency."""
-
-    async def transcribe(self, audio_wav: bytes) -> str: ...
-
-
-@dataclass(slots=True)
-class CaptureAudioClip:
-    """Internal clip payload used by service orchestration."""
-
-    filename: str | None
-    content_type: str | None
-    content: bytes
-
-
 class QuoteService:
-    """Coordinate quote domain rules with persistence and extraction."""
+    """Coordinate quote domain rules with persistence and PDF rendering."""
 
     def __init__(
         self,
         *,
         repository: QuoteRepositoryProtocol,
-        extraction_integration: ExtractionIntegrationProtocol,
-        audio_integration: AudioIntegrationProtocol,
-        transcription_integration: TranscriptionIntegrationProtocol,
         pdf_integration: PdfIntegrationProtocol,
     ) -> None:
         self._repository = repository
-        self._extraction = extraction_integration
-        self._audio = audio_integration
-        self._transcription = transcription_integration
         self._pdf = pdf_integration
-
-    async def convert_notes(self, notes: str) -> ExtractionResult:
-        """Extract structured line items from freeform notes."""
-        try:
-            return await self._extraction.extract(notes)
-        except ExtractionError as exc:
-            raise QuoteServiceError(
-                detail=f"Extraction failed: {exc}",
-                status_code=422,
-            ) from exc
-
-    async def capture_audio(self, clips: Sequence[CaptureAudioClip]) -> ExtractionResult:
-        """Normalize uploaded clips, transcribe audio, and extract quote line items."""
-        try:
-            stitched_wav = await asyncio.to_thread(
-                self._audio.normalize_and_stitch,
-                [
-                    AudioClip(
-                        filename=clip.filename,
-                        content_type=clip.content_type,
-                        content=clip.content,
-                    )
-                    for clip in clips
-                ],
-            )
-        except AudioError as exc:
-            raise QuoteServiceError(detail=str(exc), status_code=400) from exc
-
-        try:
-            transcript = await self._transcription.transcribe(stitched_wav)
-        except TranscriptionError as exc:
-            raise QuoteServiceError(
-                detail=f"Transcription failed: {exc}",
-                status_code=502,
-            ) from exc
-
-        return await self.convert_notes(transcript)
-
-    async def extract_combined(
-        self,
-        clips: Sequence[CaptureAudioClip],
-        notes: str,
-    ) -> ExtractionResult:
-        """Extract quote line items from optional clips plus optional typed notes."""
-        normalized_notes = notes.strip()
-        combined_text = normalized_notes
-
-        if clips:
-            try:
-                stitched_wav = await asyncio.to_thread(
-                    self._audio.normalize_and_stitch,
-                    [
-                        AudioClip(
-                            filename=clip.filename,
-                            content_type=clip.content_type,
-                            content=clip.content,
-                        )
-                        for clip in clips
-                    ],
-                )
-            except AudioError as exc:
-                raise QuoteServiceError(detail=str(exc), status_code=400) from exc
-
-            try:
-                transcript = await self._transcription.transcribe(stitched_wav)
-            except TranscriptionError as exc:
-                raise QuoteServiceError(
-                    detail=f"Transcription failed: {exc}",
-                    status_code=502,
-                ) from exc
-
-            combined_text = transcript
-            if normalized_notes:
-                combined_text = f"{transcript}\n\n{normalized_notes}"
-
-        if not combined_text:
-            raise QuoteServiceError(
-                detail="Provide at least one audio clip or typed notes.",
-                status_code=400,
-            )
-
-        return await self.convert_notes(combined_text)
 
     async def create_quote(self, user: User, data: QuoteCreateRequest) -> Document:
         """Create a user-owned quote and retry once on sequence collisions."""
